+++ Station.h
@@ -14,28 +14,28 @@ class Station {
 
     private:
         string name;
-        float x;
-        float y;
-        float distance;
-        float argument;
+        double x;
+        double y;
+        double distance;
+        double argument;
         int waiting_pass; // Personne en attente
         int max_pass;
         bool occupied; // Elle est occupé ?
 
     public:
-        Station(const string &name, const float &x, const float &y, const int &max_pass);
+        Station(const string &name, const double &x, const double &y, const int &max_pass);
 
         int add_waiting(const int &nb);
         void connection(Station &Other);
         void arrive();
         void leaves();
         void calculate_distance(vector<Station>& list);
-        float distance_to(Station Other);
-        float get_next_dist();
-        float get_next_arg();
+        double distance_to(Station Other);
+        double get_next_dist();
+        double get_next_arg();
         string get_name();
-        float get_x();
-        float get_y();
+        double get_x();
+        double get_y();
 
 };
